Crate skynet_substrate

Skynet Substrate-compatible SDK.

Structs

CommonOptions

Options common to all methods.

DownloadOptions

Download options.

GetEntryOptions

Get entry options.

RegistryEntry

Registry entry.

SetEntryOptions

Set entry options.

SignedRegistryEntry

Signed registry entry.

UploadOptions

Upload options.

Enums

DownloadError

Download error.

GetEntryError

Get entry error.

PinError

Pin error.

RequestError

Request error.

SetEntryError

Set entry error.

UploadError

Upload error.

Constants

DEFAULT_PORTAL_URL

The default Skynet portal URL.

HASH_LENGTH

The hash length.

PRIVATE_KEY_LENGTH

The private key length.

PUBLIC_KEY_LENGTH

The public key length.

SIGNATURE_LENGTH

The signature length.

URI_SKYNET_PREFIX

The Skynet URI protocol prefix.

Functions

download_bytes

Downloads the bytes at the given skylink.

get_entry

Gets registry entry for public_key and data_key.

get_entry_link

Gets the entry link for the entry at the given public_key and data_key. This link stays the same even if the content at the entry changes.

pin_skylink

Re-pins the given skylink.

set_data_link

Sets the datalink for the entry at the given private key and data key.

set_entry

Sets registry entry at private_key.

set_entry_data

Sets the raw entry data at the given private key and data key.

upload_bytes

Upload bytes to a file with filename.

Type Definitions

Signature

The signature type.